add debugging prints

+++ load_balance_gym/envs/load_balance_with_floodlight.py
@@ -481,9 +481,9 @@ class LoadBalanceEnv(gym.Env):
             "switch": switch_id,
             "name": flow_id,
             "priority": priority,
-            "ingress-port": numpy.int64(in_port),
+            "ingress-port": numpy.int8(in_port),
             "active": "true",
-            "actions": "output={0}".format(numpy.int64(out_port))
+            "actions": "output={0}".format(numpy.int8(out_port))
         }
 
         return json.dumps(rule)
